backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas>=2.2.0
numpy>=1.26.2
email-validator>=2.0.0
bcrypt>=4.0.0
# PostgreSQL driver (optional, only needed if using PostgreSQL)
# psycopg2-binary==2.9.9
